requirements.txt
python-dotenv
pandas
sqlalchemy
psycopg2
requests
bs4